修改格式、及数据类型

+++ SensorsAnalyticsSDK/SAConstants.m
@@ -106,7 +106,7 @@ NSString * const SA_EVENT_COMMON_OPTIONAL_PROPERTY_PROJECT = @"$project";
 NSString * const SA_EVENT_COMMON_OPTIONAL_PROPERTY_TOKEN = @"$token";
 NSString * const SA_EVENT_COMMON_OPTIONAL_PROPERTY_TIME = @"$time";
 //神策成立时间，2015-05-15 10:24:00.000，某些时间戳判断（毫秒）
-NSInteger const SA_EVENT_COMMON_OPTIONAL_PROPERTY_TIME_INT = 1431656640000;
+long const SA_EVENT_COMMON_OPTIONAL_PROPERTY_TIME_INT = 1431656640000;
 
 #pragma mark - profile
 NSString * const SA_PROFILE_SET = @"profile_set";
